+++ frontend/components/input/index.tsx
@@ -1,2 +1,3 @@
 export { default as CurrencyInput } from "./CurrencyInput";
 export { default as Input } from "./Input";
+export { default as CustomInput } from "./CustomInput";
