+++ frontend/pages/home/components/ICRC/asset/AccountElement.tsx
@@ -11,7 +11,7 @@ import { clsx } from "clsx";
 import { ChangeEvent, Fragment, useState } from "react";
 import { GeneralHook } from "../../../hooks/generalHook";
 import { toFullDecimal } from "@/utils";
-import { CustomInput } from "@components/Input";
+import { CustomInput } from "@components/input";
 import { useTranslation } from "react-i18next";
 import { useAppDispatch } from "@redux/Store";
 import { Token } from "@redux/models/TokenModels";
