+++ frontend/pages/home/components/ICRC/detail/transaction/SendForm/SenderNewContact.tsx
@@ -1,7 +1,7 @@
 import { NewContact, TransactionValidationErrorsEnum } from "@/@types/transactions";
 import { isHexadecimalValid } from "@/utils/checkers";
 import { validatePrincipal } from "@/utils/identity";
-import { CustomInput } from "@components/Input";
+import { CustomInput } from "@components/input";
 import { useAppSelector } from "@redux/Store";
 import { removeErrorAction, setErrorAction, setSenderContactNewAction } from "@redux/transaction/TransactionActions";
 import { useEffect, useState } from "react";
