+++ frontend/pages/login/components/SeedInput.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, Dispatch, SetStateAction } from "react";
 import { ReactComponent as CheckIcon } from "@assets/svg/files/edit-check.svg";
-import { CustomInput } from "@components/Input";
+import { CustomInput } from "@components/input";
 import { handleSeedAuthenticated } from "@redux/CheckAuth";
 import { clsx } from "clsx";
 import { useTranslation } from "react-i18next";
