+++ frontend/pages/login/components/WatchOnlyInput.tsx
@@ -1,6 +1,6 @@
 import { ReactComponent as CheckIcon } from "@assets/svg/files/edit-check.svg";
 //
-import { CustomInput } from "@components/Input";
+import { CustomInput } from "@components/input";
 import { decodeIcrcAccount } from "@dfinity/ledger-icrc";
 import { handlePrincipalAuthenticated } from "@redux/CheckAuth";
 import { clsx } from "clsx";
